* refactored files to proper locations and namespaces
