+++ pwiz_tools/Skyline/Model/DdaSearch/MSAmandaSearchSettings.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace pwiz.Skyline.Model.MSAmanda
+namespace pwiz.Skyline.Model.DdaSearch
 {
     public class MSAmandaSearchSettings
     {
